app/api/requirements.txt
psycopg2==2.8.6
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic==2.7.4
requests==2.32.3
python-dotenv==1.0.1
pytest==8.2.2